+++ packages/espressocash_app/lib/utils/composite_subscription_controller.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/widgets.dart';
+import 'package:rxdart/rxdart.dart';
+
+mixin CompositeSubscriptionController<T extends StatefulWidget> on State<T> {
+  final subscriptions = CompositeSubscription();
+
+  @override
+  void dispose() {
+    subscriptions.clear();
+    super.dispose();
+  }
+}